+++ docs/source/release_notes.rst
@@ -5,6 +5,7 @@ Changes
 
 Release 0.3 (October 2019)
 ----------------------------
+* Added wrapper class for different backends
 * Added documentation
 
 Release 0.2 (July 2019)